% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_calculator.R
\name{diag_phi}
\alias{diag_phi}
\title{Title}
\usage{
diag_phi(idx)
}
\arguments{
\item{idx}{The indices of precision values to select from the stored Phi^-1.}
}
\value{

}
\description{
Convenience function to create a diagonal matrix of noise precision.
Created for the edge case where only one data point is used.
}
