% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_calculator.R
\name{hatmatrix_calculator}
\alias{hatmatrix_calculator}
\title{Create a new [`HatMatrixCalculator`] object.}
\usage{
hatmatrix_calculator(Xf, Phi_inv, Xr = NULL, Sigma_inv = NULL)
}
\value{
A [`HatMatrixCalculator`] object.
}
\description{
Create a new [`HatMatrixCalculator`] object.
}
