% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_from.R
\name{hatmatrix_from.formula}
\alias{hatmatrix_from.formula}
\title{Title}
\usage{
\method{hatmatrix_from}{formula}(x, ftype, Sigma_inv, Phi_inv, i, j, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\seealso{
Other {Fitted model to hat matrix converters}: 
\code{\link{hatmatrix_from.hm_model_matrix}()},
\code{\link{hatmatrix_from.lmer}()},
\code{\link{hatmatrix_from.lm}()},
\code{\link{hatmatrix_from}()}
}
\concept{{Fitted model to hat matrix converters}}
