% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix.R
\name{hatmatrix_from.lm}
\alias{hatmatrix_from.lm}
\title{Title}
\usage{
\method{hatmatrix_from}{lm}(mod)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{
fit <- lm(mpg ~ wt, data = mtcars)
}
