% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_from.R
\name{hatmatrix_from.lmer}
\alias{hatmatrix_from.lmer}
\title{Title}
\usage{
\method{hatmatrix_from}{lmer}(mod)
}
\arguments{
\item{mod}{}
}
\value{

}
\description{
Title
}
\seealso{
Other {Fitted model to hat matrix converters}: 
\code{\link{hatmatrix_from.formula}()},
\code{\link{hatmatrix_from.hm_model_matrix}()},
\code{\link{hatmatrix_from.lm}()},
\code{\link{hatmatrix_from}()}
}
\concept{{Fitted model to hat matrix converters}}
