% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_from.R
\name{hatmatrix_from}
\alias{hatmatrix_from}
\title{Generic for obtaining the hat matrix from a fitted model}
\usage{
hatmatrix_from(x)
}
\arguments{
\item{mod}{}
}
\value{

}
\description{
Generic for obtaining the hat matrix from a fitted model
}
\seealso{
Other {Fitted model to hat matrix converters}: 
\code{\link{hatmatrix_from.formula}()},
\code{\link{hatmatrix_from.hm_model_matrix}()},
\code{\link{hatmatrix_from.lmer}()},
\code{\link{hatmatrix_from.lm}()}
}
\concept{{Fitted model to hat matrix converters}}
