% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix_calculator.R
\name{HatMatrixCalculator}
\alias{HatMatrixCalculator}
\title{R6 Class representing a HatMatrixCalculator}
\description{
A `HatMatrixCalculator` object is an [R6][R6::R6Class] object created
  by the [`hatmatrix_calculator()`] function. Internally, it stores and validates inputs to
[`hatmatrix()`] or [`hatmatrix_from()`] and contains the methods to calculate the hat matrix,
sub-matrix of the hat matrix,
or the hat matrix based on a subset of the data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{number}}{of data points (rows in \mjeqn{X_f}{X_f})}

\item{\code{rows}}{selected from \mjeqn{W}{W} in the most recent call to [$calc()]}

\item{\code{columns}}{selected from \mjeqn{W}{W} in the most recent call to [$calc()]}

\item{\code{supplied}}{design matrix for fixed effects}

\item{\code{supplied}}{design matrix for random effects}

\item{\code{supplied}}{\mjeqn{\Sigma^{-1}}{Sigma^-1}}

\item{\code{supplied}}{\mjeqn{\Phi^{-1}}{Phi^-1}}

\item{\code{hat}}{matrix from most recent call to [$calc()]}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{number}}{of data points (rows in \mjeqn{X_f}{X_f})}

\item{\code{rows}}{selected from \mjeqn{W}{W} in the most recent call to [$calc()]}

\item{\code{columns}}{selected from \mjeqn{W}{W} in the most recent call to [$calc()]}

\item{\code{supplied}}{design matrix for fixed effects}

\item{\code{supplied}}{design matrix for random effects}

\item{\code{supplied}}{\mjeqn{\Sigma^{-1}}{Sigma^-1}}

\item{\code{supplied}}{\mjeqn{\Phi^{-1}}{Phi^-1}}

\item{\code{hat}}{matrix from most recent call to [$calc()]}

\item{\code{V}}{matrix from most recent call to [$calc()]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HatMatrixCalculator$new()}}
\item \href{#method-expanded_sigma}{\code{HatMatrixCalculator$expanded_sigma()}}
\item \href{#method-diag_phi}{\code{HatMatrixCalculator$diag_phi()}}
\item \href{#method-calc}{\code{HatMatrixCalculator$calc()}}
\item \href{#method-clone}{\code{HatMatrixCalculator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HatMatrixCalculator$new(Xf, Phi_inv, Xr = NULL, Sigma_inv = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expanded_sigma"></a>}}
\if{latex}{\out{\hypertarget{method-expanded_sigma}{}}}
\subsection{Method \code{expanded_sigma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HatMatrixCalculator$expanded_sigma()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diag_phi"></a>}}
\if{latex}{\out{\hypertarget{method-diag_phi}{}}}
\subsection{Method \code{diag_phi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HatMatrixCalculator$diag_phi(idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc"></a>}}
\if{latex}{\out{\hypertarget{method-calc}{}}}
\subsection{Method \code{calc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HatMatrixCalculator$calc(
  i = 1:self$n,
  j = 1:self$n,
  return_V = FALSE,
  subset = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HatMatrixCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
