% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_model_matrix.R
\name{hmm_from.formula}
\alias{hmm_from.formula}
\title{Title}
\usage{
\method{hmm_from}{formula}(
  formula,
  family = gaussian(),
  ftype = c("lm", "glm", "lmer", "glmer", "gamm4"),
  env = parent.frame(),
  ...
)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Internal function
}
