% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix.R
\name{hatmatrix_from}
\alias{hatmatrix_from}
\title{Generic for getting the hat matrix from a fitted model object}
\usage{
hatmatrix_from(x)
}
\arguments{
\item{x}{Fitted model object}
}
\value{
The hat matrix
}
\description{
Generic for getting the hat matrix from a fitted model object
}
\concept{{hat matrix helpers}}
